+++ src/app/page.tsx
@@ -1,92 +1,273 @@
 "use client";
-
-import { useEffect, useState } from "react";
-import { apiClient } from "@/lib/apiClient";
-import { RankingType } from "@/types/ranking";
+import {Card, CardContent, CardHeader} from "@/components/ui/card";
 import Image from "next/image";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import {RankingCard} from "@/components/RankingCard";
-import {Loading} from "@/components/Loading";
-import {Button} from "@/components/ui/button";
-import Link from "next/link";
+import {Avatar, AvatarFallback, AvatarImage} from "@/components/ui/avatar";
+import {motion} from "framer-motion";
+import {CharacterCard} from "@/components/CharacterCard";
+import {getS3Url} from "@/lib/utils";
 
-export default function Home() {
-	const [isLoading, setIsLoading] = useState(true);
-	// const [rankingData, setRankingData] = useState<RankingType[]>([]);
-	const [smallAnimalRankingData, setSmallAnimalRankingData] = useState<RankingType[]>([]);
-	const [pandaRankingData, setPandaRankingData] = useState<RankingType[]>([]);
-	const [rabbitRankingData, setRabbitRankingData] = useState<RankingType[]>([]);
-	const [birdRankingData, setBirdRankingData] = useState<RankingType[]>([]);
-	const [mouseRankingData, setMouseRankingData] = useState<RankingType[]>([]);
+const TutorialPage = () => {
 
-	useEffect(() => {
-		apiClient.get("").then((res) => {
-			const data = res.data;
-			console.log(data);
-			// setSmallAnimalRankingData(data.filter((item: RankingType) => item.animal === "mouse" || item.animal === "rabbit" || item.animal === "bird").sort((a: RankingType, b: RankingType) => b.score - a.score).slice(0, 10));
-			setSmallAnimalRankingData(data.filter((item: RankingType) => item.animal !== "panda").sort((a: RankingType, b: RankingType) => b.score - a.score).slice(0, 10));
-			setPandaRankingData(data.filter((item: RankingType) => item.animal === "panda").sort((a: RankingType, b: RankingType) => b.score - a.score).slice(0, 10));
-			setRabbitRankingData(data.filter((item: RankingType) => item.animal === "rabbit").sort((a: RankingType, b: RankingType) => b.score - a.score).slice(0, 10));
-			setBirdRankingData(data.filter((item: RankingType) => item.animal === "bird").sort((a: RankingType, b: RankingType) => b.score - a.score).slice(0, 10));
-			setMouseRankingData(data.filter((item: RankingType) => item.animal === "mouse").sort((a: RankingType, b: RankingType) => b.score - a.score).slice(0, 10));
-			setIsLoading(false);
-		});
-	}, []); // 依存配列を追加して無限ループを防止
+    const infiniteAnimalVariants = {
+        jump: {
+            y: [0, -30, 0],
+            transition: {
+                duration: 1,
+                ease: "easeInOut",
+                times: [0, 0.5, 1],
+                repeat: Infinity,
+                repeatDelay: 0.2
+            }
+        },
+        fly: {
+            x: [-10, 60, -10],
+            y: [0, -50, 10, -30, 0],
+            scale: [1, 1.2, 1, 1.2, 1],
+            transition: {
+                x: {
+                    duration: 6,
+                    repeat: Infinity,
+                    ease: "easeInOut",
+                },
+                y: {
+                    duration: 3,
+                    repeat: Infinity,
+                    ease: "easeInOut",
+                },
+                scale: {
+                    duration: 0.3,
+                    repeat: Infinity,
+                    ease: "easeInOut",
+                }
+            }
+        },
+        climb: {
+            y: [0,-20,0],
+            x: [0, 5, -5, 5, -5, 0],
+            scale: [1, 1.1, 0.9, 1.1, 0.9, 1],
+            transition: {
+                y: {
+                    duration: 3,
+                    repeat: Infinity,
+                    ease: "easeInOut",
+                },
+                x: {
+                    duration: 1,
+                    repeat: Infinity,
+                    times: [0, 0.2, 0.4, 0.6, 0.8, 1],
+                },
+                scale: {
+                    duration: 4,
+                    repeat: Infinity,
+                }
+            }
+        }
+    };
 
-	return (
-		<div>
-			<div className="mb-2 p-1">
-				<h2 className="mb-4 text-center text-xl font-medium text-green-500">
-					ランキング
-				</h2>
-				<div className="mb-4 text-center">
-					<Link href="/result">
-						<Button variant="outline" size="sm" className="gap-1 bg-green-50 hover:bg-green-400 border-green-500">
-							最近の試合を見る
-						</Button>
-					</Link>
-				</div>
-			</div>
+    return (
+        <div className={"flex flex-col justify-center items-center"}>
+            <h1 className={"text-4xl text-green-500 font-medium"}>チュートリアル</h1>
+
+            <h2 className={"my-10 text-2xl text-green-500 font-medium"}>ストーリー</h2>
+            <Card className={"sm:w-4xl mx-3"}>
+                <CardHeader>
+                    <div className={"flex justify-center text-2xl text-green-500 font-medium"}>
+                        パンダから逃げろ！
+                    </div>
+                </CardHeader>
+                <CardContent className={"flex flex-col justify-center gap-6"}>
+                    <Image src={getS3Url("/pandator-scene.png")} alt={"pandator-scene"} width={800} height={200} className={"mx-auto"}/>
+                    <div className={"text-center text-xl"}>ある館での出来事...</div>
+                    <div
+                        className={"text-center text-xl"}>小動物達はパンダに捕まってしまい、檻に閉じこめられてしまった。
+                    </div>
+                    <div
+                        className={"text-center text-xl"}>なんとか小動物達は檻を脱出したが、帰宅したパンダに見つかってしまう...
+                    </div>
+                    <div className={"text-center text-xl"}>果たして小動物達はパンダから逃げ切れるのか!?</div>
+                </CardContent>
+            </Card>
+
+            <h2 className={"my-10 text-2xl text-green-500 font-medium"}>基本ルール</h2>
+            <Card　className={"sm:w-4xl mx-3 px-2 flex flex-col justify-center items-center"}>
+                <div className={"text-2xl font-bold"}>パンダ</div>
+                <div>パンダは、逃げ回る小動物達を銃で網を放ち、小動物を捕まえましょう！</div>
+                <div className={"text-2xl font-bold"}>小動物</div>
+                <div>小動物達は、パンダから捕まらないように逃げ切りましょう！</div>
+                <div>また、他の小動物へ妨害をして、自分が逃げ切りましょう！</div>
+            </Card>
 
 
-			{isLoading && (
-				<Loading/>
-			)}
+            <h2 className={"my-10 text-2xl text-green-500 font-medium"}>動物毎の操作方法</h2>
+            <div className={"gap-6"}>
+                <CharacterCard animal={"panda"}>
+                    <div className={"flex flex-col items-center gap-6"}>
+                        <div className={"text-2xl font-bold"}>キャラクター選択</div>
+                        <div>銃に接続されているAとBボタンを押してください。</div>
+                        <Image src={getS3Url("/tutorial/game/select.png")} alt={"キャラクター選択画面"} width={400} height={400}/>
 
-			<Tabs defaultValue="all" className="w-xs sm:w-2xl mx-auto">
-				<TabsList className="grid w-full grid-cols-5">
-				<TabsTrigger value="all">小動物全体</TabsTrigger>
-					<TabsTrigger value="panda">パンダ</TabsTrigger>
-					<TabsTrigger value="rabbit">うさぎ</TabsTrigger>
-					<TabsTrigger value="bird">鳥</TabsTrigger>
-					<TabsTrigger value="mouse">ねずみ</TabsTrigger>
-				</TabsList>
-				<TabsContent value="all">
-					<RankingCard rankingData={smallAnimalRankingData} />
-				</TabsContent>
-				<TabsContent value="panda">
-					<RankingCard rankingData={pandaRankingData}/>
-				</TabsContent>
-				<TabsContent value="rabbit">
-					<RankingCard rankingData={rabbitRankingData}/>
-				</TabsContent>
-				<TabsContent value="bird">
-					<RankingCard rankingData={birdRankingData}/>
-				</TabsContent>
-				<TabsContent value="mouse">
-					<RankingCard rankingData={mouseRankingData}/>
-				</TabsContent>
-			</Tabs>
+                    </div>
+                    <div className={"my-8 flex flex-col  items-center gap-6"}>
+                        <div className={"text-2xl font-bold"}>パンダの操作</div>
+                        <div>銃の引き金を引くと網が発射！</div>
+                        <Image src={getS3Url("/tutorial/game/PandaShotGame.gif")} alt={"網が出ている様子"} width={400} height={400}/>
+                        {/*<div>小動物が網の中に入るように、狙い撃て!</div>*/}
+                        {/*<Image src="/RabbitReal.gif" alt={"動物が捕まっている様子"} width={400} height={400}/>*/}
 
-			{!isLoading&& (
-				<div className="mt-6 flex animate-bounce items-center justify-center text-2xl">
-					<span className="mr-2">🐼</span>
-					<span className="mr-2">🐰</span>
-					<span className="mr-2">🐦️</span>
-					<span>🐭</span>
-				</div>
-			)}
+                    </div>
 
-		</div>
-	);
+                </CharacterCard>
+                <CharacterCard animal={"rabbit"}>
+                    <div className={"flex flex-col items-center gap-6"}>
+                        <div className={"text-2xl font-bold"}>キャラクター選択</div>
+                        <div>選択画面からRabbitを選んでください。</div>
+                        <Image src={getS3Url("/tutorial/game/select.png")} alt={"ウサギを選択してる画像"} width={400} height={400}/>
+                    </div>
+
+                    <div className={"my-8 flex flex-col  items-center gap-6"}>
+                        <div className={"text-2xl font-bold"}>ウサギの操作</div>
+                        <div className={"flex justify-center gap-6"}>
+                            <div>腕を振って走ろう!</div>
+                        </div>
+                        <div className={"flex"}>
+                            <div className={"mx-auto flex flex-col gap-1"}>
+                                <Image src={getS3Url("/tutorial/real/WalkAction.gif")} alt={"腕を振ってる画像"} width={400} height={400}/>
+                                {/*<Image src="/RabbitReal.gif" alt={"ゲーム内でジャンプしてる画像"} width={400}*/}
+                                {/*       height={400}/>*/}
+                            </div>
+                        </div>
+
+                        <div className={"flex justify-center gap-6"}>
+                            <div>Aボタンを押しながら腕を大きく振ってジャンプ!</div>
+                            <motion.div
+                                variants={infiniteAnimalVariants}
+                                animate="jump"
+                            >
+                                <span className={"text-2xl"}>🐇</span>
+                            </motion.div>
+                        </div>
+                        <div className={"flex relative"}>
+                            <div className={"mx-auto flex flex-col gap-1"}>
+                                <Image src={getS3Url("/tutorial/real/RabbitReal.gif")} alt={"real"} width={400} height={400}/>
+                                {/*TODO: rabbitのゲーム動画*/}
+                                {/*<Image src="/RabbitReal.gif" alt={"game"} width={400} height={400}/>*/}
+
+                            </div>
+                        </div>
+
+                    </div>
+                </CharacterCard>
+                <CharacterCard animal={"bird"}>
+                    <div className={"flex flex-col items-center gap-6"}>
+                        <div className={"text-2xl font-bold"}>キャラクター選択</div>
+                        <div>選択画面からBirdを選んでください。</div>
+                        <Image src={getS3Url("/tutorial/game/select.png")} alt={"鳥を選択してる画像"} width={400} height={400}/>
+                    </div>
+                    <div className={"my-8 flex flex-col  items-center gap-6"}>
+                        <div className={"text-2xl font-bold"}>鳥の操作</div>
+                        <div className={"flex justify-center gap-6"}>
+                            <div>右スティックで歩こう！</div>
+                        </div>
+                        <div className={"flex relative"}>
+                            <div className={"mx-auto flex gap-1"}>
+                                <Image src={getS3Url("/tutorial/real/WalkAction.gif")} alt={"ゲーム内で歩いてる画像"} width={400}
+                                       height={400}/>
+                            </div>
+                        </div>
+
+                        <div className={"flex justify-center gap-6"}>
+                            <div>腕をバタバタさせて空を飛ぼう!</div>
+                            <motion.div
+                                variants={infiniteAnimalVariants}
+                                animate="fly"
+                            >
+                                <span className={"text-2xl"}>🕊️</span>
+                            </motion.div>
+                        </div>
+                        <div className={"flex relative"}>
+                            <div className={"mx-auto flex flex-col gap-3"}>
+                                <Image src={getS3Url("/tutorial/real/BirdFlyReal.gif")} alt={"腕をバタバタさせてる様子"} width={400} height={400}/>
+                                <Image src={getS3Url("/tutorial/game/BirdFly.gif")} alt={"鳥が飛んでる画像"} width={400} height={400}/>
+                            </div>
+                        </div>
+                    </div>
+                </CharacterCard>
+                <CharacterCard animal={"mouse"}>
+                    <div className={"flex flex-col items-center gap-6"}>
+                        <div className={"text-2xl font-bold"}>キャラクター選択</div>
+                        <div>選択画面からMouseを選んでください。</div>
+                        <Image src={getS3Url("/tutorial/game/select.png")} alt={"ネズミを選択してる画像"} width={400} height={400}/>
+                    </div>
+                    <div className={"my-8 flex flex-col  items-center gap-6"}>
+                        <div className={"text-2xl font-bold"}>ネズミの操作</div>
+                        <div className={"flex justify-center gap-6"}>
+                            <div>腕を振って歩こう！</div>
+                        </div>
+                        <div className={"flex relative"}>
+                            <div className={"mx-auto flex gap-1"}>
+                                <Image src={getS3Url("/tutorial/real/WalkAction.gif")} alt={"腕を振ってる画像"} width={400} height={400}/>
+                            </div>
+                        </div>
+
+                        <div className={"flex justify-center gap-6"}>
+                            <div>壁に向かって腕を振って壁を登ろう!</div>
+                            <motion.div
+                                variants={infiniteAnimalVariants}
+                                animate="climb"
+                            >
+                                <span className={"text-2xl"}>🐀</span>
+                            </motion.div>
+                        </div>
+                        <div className={"flex relative"}>
+                            <div className={"mx-auto flex flex-col gap-1"}>
+                                <Image src={getS3Url("/tutorial/real/MouseClimbReal.gif")} alt={"腕を振っている様子"} width={400} height={400}/>
+                                <Image src={getS3Url("/tutorial/game/MouseClimb.gif")} alt={"ネズミが壁を登っている様子"} width={400} height={400}/>
+                            </div>
+                        </div>
+                    </div>
+                </CharacterCard>
+            </div>
+
+            {/*<Card className={"w-4xl flex flex-col p-0 border-green-200 border-"}>*/
+            }
+            {/*    <CardHeader className={"bg-blue-50  items-center"}>*/
+            }
+            {/*        <div className="flex">*/
+            }
+            {/*            <Avatar className={"size-10"}>*/
+            }
+            {/*                <AvatarImage src={"rabbit.png"}/>*/
+            }
+            {/*                <AvatarFallback>CN</AvatarFallback>*/
+            }
+            {/*            </Avatar>*/
+            }
+            {/*            <div className={"text-4xl"}>Rabbit</div>*/
+            }
+            {/*        </div>*/
+            }
+            {/*    </CardHeader>*/
+            }
+            {/*    <div className={"flex relative"}>*/
+            }
+            {/*        <div className={"mx-auto flex gap-1"}>*/}
+            {/*            <Image src="/RabbitReal.gif" alt={"real"} width={400} height={400}/>*/}
+            {/*            <Image src="/RabbitReal.gif" alt={"game"} width={400} height={400}/>*/}
+            {/*        </div>*/}
+            {/*    </div>*/}
+            {/*    <div className={"flex justify-center gap-6"}>*/}
+            {/*        <h3>腕を大きく振ってジャンプ!</h3>*/}
+            {/*        <motion.div*/}
+            {/*            variants={infiniteJumpVariants}*/}
+            {/*            animate="jump"*/}
+            {/*        >*/}
+            {/*            <span className={"text-2xl"}>🐇</span>*/}
+            {/*        </motion.div>*/}
+            {/*    </div>*/}
+
+            {/*</Card>*/}
+        </div>
+    )
 }
+
+export default TutorialPage;
